AI_PowerPoint_Generator/requirements.txt
flask
flask-sqlalchemy
flask-login
flask-wtf
google-genai
python-pptx
requests
python-dotenv
passlib
email-validator
gunicorn
psycopg2-binary
pytest
werkzeug
